virtio-pci: make pci bars configurable
Add msix_bar and modern_mem_bar fields to VirtIOPCIProxy.  They can be
used to configure which pci regions are used for the virtio 1.0 memory
bar and the msix bar.

For legacy/transitional devices the legacy bar is region 0 and the msix
bar is region 1.  Only the modern bar can be configured, and it must be
2 or larger.  Default is 2.

For legacy-free devices the modern bar is region 0 by default and the
msix bar is 2 by default.

Use case: For VirtIOPCIProxy subclasses which need additional pci bars,
such as virtio-vga.  With the new fields they can make sure the regions
do not conflict.

Signed-off-by: Gerd Hoffmann <kraxel@redhat.com>